div done

+++ devision_mahima.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 void devision_mahima(int a,int b)
 {
-	float c=a/b;
-	printf("the division value is %f",c);
+        int c=a/b;
+        printf("the division value is %d",c);
 }
